#6 | remove file schema from store; create model files in core module
