+++ src/app/core/models/cart.model.ts
@@ -1,0 +1,5 @@
+export interface Cart {
+  id: string;
+  clientId: string;
+  products: string[];
+}
