+++ src/app/core/models/category.model.ts
@@ -1,0 +1,5 @@
+export interface Category {
+  id: string;
+  name: string;
+  createdAt: number;
+}
