+++ src/app/core/models/chat.model.ts
@@ -1,0 +1,12 @@
+export interface SupportChat {
+  id: string;
+  clientId: string;
+  messages: SupportMessage[];
+  createdAt: number;
+}
+
+export interface SupportMessage {
+  userId: string;
+  text: string;
+  createdAt: number;
+}
