+++ src/app/core/models/index.ts
@@ -1,0 +1,6 @@
+export * from './cart.model';
+export * from './category.model';
+export * from './chat.model';
+export * from './order.model';
+export * from './product.model';
+export * from './user.model';
