+++ src/app/core/models/order.model.ts
@@ -1,0 +1,11 @@
+import { OrderStatus } from "../enums/order-status.enum";
+
+export interface Order {
+  id: string;
+  orderNumber: number;
+  clientId: string;
+  products: string[];
+  total: number;
+  status: OrderStatus;
+  createdAt: number;
+}
