+++ src/app/core/models/product.model.ts
@@ -1,0 +1,9 @@
+export interface Product {
+  id: string;
+  name: string;
+  description: string;
+  price: number;
+  photo?: string;
+  categoryId: string;
+  createdAt: number;
+}
