+++ src/app/core/services/auth.service.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { User } from '../store/schemas/app.schema';
+import { User } from '../models';
 
 interface IRegistrCreds {
   username: string;
