+++ src/app/core/store/state/app.state.ts
@@ -1,5 +1,6 @@
 import { UserRole } from '../../enums/user-role.enum';
-import { Category, Product, User } from '../schemas/app.schema';
+import { Category, Product, User } from '../../models';
+
 export interface IAppState {
   user: User;
   product?: Product;
