+++ src/app/static/auth/auth.module.ts
@@ -6,6 +6,7 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 const routes: Routes = [
   { path: '', component: AuthComponent },
+  { path: '', redirectTo: '/' }
 ]
 
 @NgModule({
